First code review changes
Sets correct code style.
Adds factory to CompositeExtractor.
Sets correct supportsJoin() and requires128BitTraceId().
Extracts XCloudTraceContextExtractor to its own class.
Sets brave version in dependencyManagement.
