+++ pom.xml
@@ -58,6 +58,7 @@
     <!-- must match https://github.com/grpc/grpc-java/blob/master/SECURITY.md#netty -->
     <grpc.version>1.11.0</grpc.version>
     <netty-tcnative-boringssl-static.version>2.0.7.Final</netty-tcnative-boringssl-static.version>
+    <brave.version>5.0.0</brave.version>
   </properties>
 
   <name>Zipkin Google Cloud Platform(Parent)</name>
@@ -156,6 +157,12 @@
         <groupId>org.hamcrest</groupId>
         <artifactId>hamcrest-core</artifactId>
         <version>1.3</version>
+      </dependency>
+
+      <dependency>
+        <groupId>io.zipkin.brave</groupId>
+        <artifactId>brave</artifactId>
+        <version>5.0.0</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
