+++ propagation-stackdriver/pom.xml
@@ -20,11 +20,11 @@
     <parent>
         <artifactId>zipkin-gcp-parent</artifactId>
         <groupId>io.zipkin.gcp</groupId>
-        <version>0.4.2-SNAPSHOT</version>
+        <version>0.5.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>zipkin-propagation-stackdriver</artifactId>
+    <artifactId>brave-propagation-stackdriver</artifactId>
 
     <properties>
         <main.basedir>${project.basedir}/..</main.basedir>
@@ -34,7 +34,6 @@
         <dependency>
             <groupId>io.zipkin.brave</groupId>
             <artifactId>brave</artifactId>
-            <version>5.0.0</version>
         </dependency>
     </dependencies>
 </project>
